Made minor fixes: code splitting, indentation fix, etc.
