+++ src/assets/styles/icon.sass
@@ -1,0 +1,6 @@
+.icon
+  display: flex
+  width: 20px
+  height: 20px
+  color: tomato
+  cursor: pointer
