+++ src/components/TodoList.vue
@@ -16,18 +16,39 @@
 
 <script>
 import TodoItem from "@/components/TodoItem.vue";
-import { mapGetters } from "vuex";
+import { mapGetters, mapActions } from "vuex";
 
 export default {
   name: "TodoList",
+
   components: {
     TodoItem,
   },
+
   computed: {
     ...mapGetters({
       items: "filteredItemsByDate",
     }),
     ...mapGetters("database", ["areItemsLoaded"]),
+  },
+
+  methods: {
+    ...mapActions(["loadItems"]),
+    ...mapActions("database", ["getFromDatabase"]),
+  },
+
+  mounted() {
+    let date = new Date();
+    const year = date.getFullYear();
+    const month = date.getMonth();
+    const startDate = new Date(year, month, 1);
+    const endDate =
+      month === 11 ? new Date(year + 1, 0, 1) : new Date(year, month + 1, 1);
+    const payload = {
+      startDate,
+      endDate,
+    };
+    this.getFromDatabase(payload);
   },
 };
 </script>
